requirements.txt
discord.py
asyncio
datetime
itertools
os
json
random
discord.ext
math
discord.ui
PIL
aiohttp
time
io
urllib.request
matplotlib.pyplot
matplotlib.image
matplotlib
timedelta
requests
yt-dlp>=2024.01.01
python-dotenv
functools